Allow Rails to improve performance for simple associations
ActiveRecord has a performance optimisation which allows it to
automatically detect the inverse of an association, i.e.:

  foo = Foo.find(1)  # performs DB call
  foo.bar            # performs DB call
  foo.bar.foo        # no additional call!

This feature was added in Rails 4.2 [1]. However, if you specify a
foreign_key, this behaviour is skipped which means that the third
expression above will make a DB call. Since the foreign_key is only
necessary when it doesn't match the Rails' expectation (`{resource}_id`)
we can just omit the option in those cases, which means lookups will
sometimes be faster again.

[1]: https://github.com/rails/rails/pull/10886